refactor: pomodoro file structure
